Allow users to set properties on list nodes
This will allow users to set custom properties on list nodes. An
example of when this could be useful: if you would like to render
an `ol` tag with `type="A"` to generate an ordered list using
alphabetic characters instead of numbers. Currently, there is no
way for a node to remember what type of ordered list it should be.
With this change, you could save the type onto the node in `data`
and then read that off the node when rendering to decide how you
will render the block (assuming you are using a custom renderer).